PhoneLogin/requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
authy==2.2.3
Babel==2.7.0
backcall==0.1.0
beautifulsoup4==4.8.0
Brlapi==0.6.6
certifi==2019.9.11
chardet==3.0.4
Click==7.0
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.4
cupshelpers==1.0
decorator==4.1.2
defer==1.0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
Django==2.2.5
django-allauth==0.40.0
django-bootstrap4==1.0.0
django-braces==1.13.0
django-cors-headers==3.0.2
django-oauth-toolkit==1.2.0
django-phone-field==1.8.0
django-phone-login==1.1.3
django-phonenumber-field==3.0.1
django-rest-auth==0.9.5
django-sendsms==0.2.3
django-twilio==0.10.0
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
env==0.1.0
Flask==1.1.1
httplib2==0.9.2
idna==2.8
ipython==5.5.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
itypes==1.1.0
jedi==0.15.1
Jinja2==2.10.1
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
mysqlclient==1.3.10
netifaces==0.10.4
numpy==1.17.0
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
parso==0.5.1
pexpect==4.2.1
phonenumbers==8.10.16
phonenumberslite==8.10.18
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
protobuf==3.0.0
psycopg2-binary==2.8.3
ptyprocess==0.6.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.4.2
pygobject==3.26.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==0.8.0
PyNaCl==1.1.2
pyperclip==1.7.0
pyRFC3339==1.0
PySocks==1.7.0
python-apt==1.6.2
python-debian==0.1.32
python-dotenv==0.10.3
python-memcached==1.59
python3-openid==3.1.0
pytz==2019.2
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.22.0
requests-oauthlib==1.2.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplegeneric==0.8.1
simplejson==3.16.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.3
sqlparse==0.3.0
system-service==0.3
systemd-python==234
traitlets==4.3.2
twilio==6.30.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
uritemplate==3.0.0
urllib3==1.25.3
usb-creator==0.3.3
virtualenv==16.7.5
wadllib==1.3.2
wcwidth==0.1.7
Werkzeug==0.15.6
xkit==0.0.0
zope.interface==4.3.2
